+++ pkg/analysis_server/lib/src/analysis_server_abstract.dart
@@ -78,7 +78,6 @@ abstract class AbstractAnalysisServer {
     '**/*.${AnalysisEngine.SUFFIX_DART}',
     '**/*.${AnalysisEngine.SUFFIX_HTML}',
     '**/*.${AnalysisEngine.SUFFIX_HTM}',
-    '**/${AnalysisEngine.ANALYSIS_OPTIONS_FILE}',
     '**/${AnalysisEngine.ANALYSIS_OPTIONS_YAML_FILE}',
     '**/${AnalysisEngine.PUBSPEC_YAML_FILE}',
     '**/${AnalysisEngine.ANDROID_MANIFEST_FILE}'