+++ pkg/analysis_server/test/integration/analysis/analysis_options_test.dart
@@ -23,7 +23,7 @@ class OptionsIntegrationTest extends AbstractAnalysisServerIntegrationTest {
     standardAnalysisSetup();
   }
 
-  Future<void> test_option_warning_newOptionFile() async {
+  Future<void> test_option_warning_optionFile() async {
     String options = sourcePath(AnalysisEngine.ANALYSIS_OPTIONS_YAML_FILE);
     writeFile(options, '''
 linter:
@@ -46,32 +46,5 @@ linter:
     expect(error.location.length, 'camel_case_typo'.length);
     expect(error.location.startLine, 3);
     expect(error.location.startColumn, 7);
-  }
-
-  Future<void> test_option_warning_oldOptionFile() async {
-    String options = sourcePath(AnalysisEngine.ANALYSIS_OPTIONS_FILE);
-    writeFile(options, '''
-linter:
-  rules:
-    - camel_case_types
-''');
-
-    optionsAnalysisSetup();
-
-    await analysisFinished;
-
-    expect(currentAnalysisErrors[options], isList);
-    List<AnalysisError> errors = currentAnalysisErrors[options];
-    expect(errors, hasLength(1));
-    AnalysisError error = errors[0];
-    expect(error.location.file, options);
-    expect(error.severity, AnalysisErrorSeverity.INFO);
-    expect(error.type, AnalysisErrorType.HINT);
-    expect(error.location.offset, 0);
-    expect(error.location.length, 1);
-    expect(error.location.startLine, 1);
-    expect(error.location.startColumn, 1);
-    expect(error.message,
-        'The name of the analysis options file .analysis_options is deprecated; consider renaming it to analysis_options.yaml.');
   }
 }